Carry out gdrive related procedures from converters module
