+++ src/ggrc/views/converters.py
@@ -9,8 +9,6 @@ including the import/export api endponts.
 
 from logging import getLogger
 
-from apiclient import discovery
-from apiclient import http
 from apiclient.errors import HttpError
 
 from flask import current_app
@@ -22,12 +20,11 @@ from werkzeug.exceptions import (
 )
 
 from ggrc import settings
-from ggrc_gdrive_integration import get_http_auth
 from ggrc_gdrive_integration import verify_credentials
+from ggrc_gdrive_integration import file_actions as fa
 from ggrc.app import app
 from ggrc.converters.base import Converter
 from ggrc.converters.import_helper import generate_csv_string
-from ggrc.converters.import_helper import read_csv_file
 from ggrc.query.exceptions import BadQueryException
 from ggrc.query.builder import QueryHelper
 from ggrc.login import login_required
@@ -62,23 +59,6 @@ def parse_export_request():
   return request.json
 
 
-def create_gdrive_file(csv_string, filename):
-  """Post text/csv data to a gdrive file"""
-  http_auth = get_http_auth()
-  drive_service = discovery.build('drive', 'v3', http=http_auth)
-  # make export to sheets
-  file_metadata = {
-      'name': filename,
-      'mimeType': 'application/vnd.google-apps.spreadsheet'
-  }
-  media = http.MediaInMemoryUpload(csv_string,
-                                   mimetype='text/csv',
-                                   resumable=True)
-  return drive_service.files().create(body=file_metadata,
-                                      media_body=media,
-                                      fields='id, name, parents').execute()
-
-
 def handle_export_request():
   """Export request handler"""
   try:
@@ -97,7 +77,7 @@ def handle_export_request():
       object_names = "_".join(converter.get_object_names())
       filename = "{}.csv".format(object_names)
       if export_to == "gdrive":
-        gfile = create_gdrive_file(csv_string, filename)
+        gfile = fa.create_gdrive_file(csv_string, filename)
         headers = [('Content-Type', 'application/json'), ]
         return current_app.make_response((json.dumps(gfile), 200, headers))
       if export_to == "csv":
@@ -144,39 +124,10 @@ def parse_import_request():
     raise BadRequest("Export failed due incorrect request data.")
 
 
-def get_gdrive_file(file_data):
-  """Get text/csv data from gdrive file"""
-  http_auth = get_http_auth()
-  try:
-    drive_service = discovery.build('drive', 'v3', http=http_auth)
-    # check file type
-    file_meta = drive_service.files().get(fileId=file_data['id']).execute()
-    if file_meta.get("mimeType") == "text/csv":
-      file_data = drive_service.files().get_media(
-          fileId=file_data['id']).execute()
-    else:
-      file_data = drive_service.files().export_media(
-          fileId=file_data['id'], mimeType='text/csv').execute()
-    csv_data = read_csv_file(file_data.splitlines())
-  except AttributeError:
-    # when file_data has no splitlines() method
-    raise BadRequest("Wrong file type.")
-  except HttpError as e:
-    message = json.loads(e.content).get("error").get("message")
-    if e.resp.code == 404:
-      raise NotFound(message)
-    if e.resp.code == 401:
-      raise Unauthorized("{} Try to reload /import page".format(message))
-    raise InternalServerError(message)
-  except:  # pylint: disable=bare-except
-    raise InternalServerError("Import failed due to internal server error.")
-  return csv_data
-
-
 def handle_import_request():
   """Import request handler"""
   dry_run, file_data = parse_import_request()
-  csv_data = get_gdrive_file(file_data)
+  csv_data = fa.get_gdrive_file(file_data)
   try:
     converter = Converter(dry_run=dry_run, csv_data=csv_data)
     converter.import_csv()
